Update meta tags for mirror and poaps
