+++ src/components/Heads/index.js
@@ -1,0 +1,5 @@
+import PostHead from './PostHead';
+
+export default {
+  PostHead
+};
