+++ src/pages/post/[id].js
@@ -14,14 +14,7 @@ import { usePost } from '../../hooks/queries';
 import withSuspense from '../../hoc/withSuspense';
 import { LOADER_TYPE, REACT_QUERY_KEYS } from '../../constants/enum';
 import callYupApi from '../../apis/base_api';
-import PostCard from '../../components/PostCard/PostCard';
-import {
-  getAbsolutePath,
-  isWeb3Post,
-  summarizeContent
-} from '../../utils/helpers';
-import { COMPANY_NAME } from '../../constants/const';
-import YupHead from '../../components/YupHead';
+import PostHead from '../../components/Heads/PostHead';
 
 function PostDetails() {
   const router = useRouter();
@@ -29,51 +22,9 @@ function PostDetails() {
   const { id } = router.query;
   const post = usePost(id);
 
-  const isWeb3 = isWeb3Post(post.tag);
-  const metaOg = {};
-  const metaTwitter = {};
-  const metaOther = {};
-  let metaTitle;
-  let metaDescription;
-  let metaImage;
-
-  if (isWeb3) {
-    const { web3Preview = {} } = post;
-    const { attachments = [], linkPreview = [] } = web3Preview;
-    const author =
-      web3Preview.creator?.fullname || web3Preview.meta?.displayName;
-
-    metaTitle = `Post by ${author} | ${COMPANY_NAME}`;
-    metaDescription = `${summarizeContent(web3Preview.content)}`;
-
-    if (attachments.length > 0 && attachments[0].images?.length > 0) {
-      [metaImage] = attachments[0].images;
-    } else if (linkPreview.length > 0) {
-      metaImage = linkPreview[0].img;
-    }
-
-    metaImage ||= getAbsolutePath('/images/metaImages/main-meta.jpg');
-
-    metaOg.author = author;
-    metaOg.url = getAbsolutePath(`/post/${id}`);
-    metaOg.site_name = COMPANY_NAME;
-    metaOg.type = 'article';
-
-    metaTwitter.card = 'summary_large_image';
-
-    metaOther.author = author;
-  }
-
   return (
     <ErrorBoundary>
-      <YupHead
-        title={metaTitle}
-        description={metaDescription}
-        image={metaImage}
-        metaOg={metaOg}
-        metaOther={metaOther}
-        metaTwitter={metaTwitter}
-      />
+      <PostHead post={post} />
       <YupPageWrapper>
         <YupPageHeader scrolled={windowScrolled} noborder>
           <YupContainer>
